feat(SideBar): add media query
