+++ packages/hedgehog-lab/src/components/Layout/RLayoutStates.ts
@@ -1,0 +1,6 @@
+import {atom} from "recoil";
+
+export const sideBarOpenState = atom({
+    key: 'sideBarOpen',
+    default: false
+})
