+++ packages/hedgehog-lab/src/components/YourCode/Config/SideBar.ts
@@ -1,0 +1,1 @@
+export const sideBarWidth = 240;
